moved SQLModel models
