+++ purch/auth/router.py
@@ -4,7 +4,7 @@ from typing import Annotated
 from fastapi import APIRouter, Depends, HTTPException, status
 from fastapi.security import OAuth2PasswordRequestForm
 
-from purch.utils.config import get_settings, Settings
+from purch.common.config import get_settings, Settings
 from purch.auth.response_models import Token
 from purch.auth.security import (
     create_purch_jwt_access_token,
