+++ purch/core/__init__.py
@@ -1,7 +1,1 @@
-import taskiq_fastapi
 
-from purch.core.taskiq import setup_taskiq_broker_and_scheduler
-
-broker, scheduler = setup_taskiq_broker_and_scheduler()
-
-taskiq_fastapi.init(broker, "purch.main:app")
