--- purch/core/finance.py
@@ -1,6 +1,0 @@
-from enum import StrEnum, auto
-
-
-class PlaidEnvs(StrEnum):
-    sandbox = auto()
-    production = auto()
