+++ purch/finance/router.py
@@ -10,7 +10,7 @@ from purch.finance.tokens import get_plaid_link_token, get_plaid_access_token
 from purch.finance.response_models import LinkTokenResponse
 from purch.finance.tasks import create_and_store_item_and_accounts, sync_transactions
 from purch.auth.security import get_current_active_user
-from purch.utils.config import Settings, get_settings
+from purch.common.config import Settings, get_settings
 
 router = APIRouter(dependencies=[Depends(get_current_active_user)])
